requirements.txt
streamlit
pandas
numpy
scikit-learn
shap
lime
matplotlib
seaborn
plotly
joblib
python-dotenv
openai
requests